leejichan/ezenbucks_api/requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.11.0
flake8==3.8.2
mccabe==0.6.1
mysqlclient @ file:///C:/mysqlclient-1.4.6-cp38-cp38-win_amd64.whl
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.3.1
